Generalize set_literal to push_memo instruction.
We are transitioning away from using a literals Tuple and that requires that
we do not alias indexes in the literals Tuple like we do now with set_literal
instruction.

The literal values are inline with the instructions, so the push_memo
instruction enables a way to compute a value at runtime and store it in the
instruction stream.  This permits better safety by not allowing arbitrary
instruction stream locations to be overwritten.

The semantics of push_memo are the following:

instruction push_memo(literal) [ value -- memo ]

If the stack_top() value is nil, the stack_top() is replaced with the memo
value. If the stack_top() value is not nil, the value is memoized (ie the
location of the literal in the instruction stream is replaced with the
stack_top() value. The memoize value is initialized to nil.