Instruction set cleanup.
Removed the meta_X instructions. These will be replaced with more specific
instructions that are not 'short-cut optimizations' per se but are rather just
optimized for specific cases.

Removed the giX aliases as they are not as useful as the explicit instruction
names such as goto_if_true, etc.

Cleaned up the {push, find}_const instructions.

Generalized the 'push' instruction to handle all literal values. Removed the
push_int instruction because all literals will be inserted inline when the
CompiledCode instance is specialized to a MachineCode instance for execution.